+++ packages/stream_chat/lib/src/core/models/message.dart
@@ -11,6 +11,7 @@ import 'package:stream_chat/src/core/models/poll.dart';
 import 'package:stream_chat/src/core/models/reaction.dart';
 import 'package:stream_chat/src/core/models/reaction_group.dart';
 import 'package:stream_chat/src/core/models/user.dart';
+import 'package:stream_chat/src/core/util/extension.dart';
 import 'package:stream_chat/src/core/util/serializer.dart';
 import 'package:uuid/uuid.dart';
 
@@ -69,6 +70,7 @@ class Message extends Equatable implements ComparableFieldProvider {
     this.moderation,
     this.draft,
     this.reminder,
+    this.channelRole,
   })  : id = id ?? const Uuid().v4(),
         type = MessageType(type),
         pinExpires = pinExpires?.toUtc(),
@@ -314,6 +316,19 @@ class Message extends Equatable implements ComparableFieldProvider {
   /// This is present when a user has set a reminder for this message.
   final MessageReminder? reminder;
 
+  static Object? _channelRoleReadValue(Map<Object?, Object?> json, String key) {
+    // Extract the channel role from the member object if present.
+    final member = json['member'];
+    if (member is! Map<String, Object?>) return null;
+
+    final channelRole = member[key].safeCast<String>();
+    return channelRole;
+  }
+
+  /// The role of the user in the channel the message belongs to.
+  @JsonKey(includeToJson: false, readValue: _channelRoleReadValue)
+  final String? channelRole;
+
   /// Message custom extraData.
   final Map<String, Object?> extraData;
 
@@ -362,6 +377,7 @@ class Message extends Equatable implements ComparableFieldProvider {
     'moderation_details',
     'draft',
     'reminder',
+    'member',
   ];
 
   /// Serialize to json.
@@ -424,6 +440,7 @@ class Message extends Equatable implements ComparableFieldProvider {
     Moderation? moderation,
     Object? draft = _nullConst,
     Object? reminder = _nullConst,
+    String? channelRole,
   }) {
     assert(() {
       if (pinExpires is! DateTime &&
@@ -506,6 +523,7 @@ class Message extends Equatable implements ComparableFieldProvider {
       draft: draft == _nullConst ? this.draft : draft as Draft?,
       reminder:
           reminder == _nullConst ? this.reminder : reminder as MessageReminder?,
+      channelRole: channelRole ?? this.channelRole,
     );
   }
 
@@ -551,6 +569,7 @@ class Message extends Equatable implements ComparableFieldProvider {
       moderation: other.moderation,
       draft: other.draft,
       reminder: other.reminder,
+      channelRole: other.channelRole,
     );
   }
 
@@ -616,6 +635,7 @@ class Message extends Equatable implements ComparableFieldProvider {
         moderation,
         draft,
         reminder,
+        channelRole,
       ];
 
   @override