+++ packages/stream_chat_persistence/lib/src/db/drift_chat_database.g.dart
@@ -842,6 +842,12 @@ class $MessagesTable extends Messages
   late final GeneratedColumn<String> userId = GeneratedColumn<String>(
       'user_id', aliasedName, true,
       type: DriftSqlType.string, requiredDuringInsert: false);
+  static const VerificationMeta _channelRoleMeta =
+      const VerificationMeta('channelRole');
+  @override
+  late final GeneratedColumn<String> channelRole = GeneratedColumn<String>(
+      'channel_role', aliasedName, true,
+      type: DriftSqlType.string, requiredDuringInsert: false);
   static const VerificationMeta _pinnedMeta = const VerificationMeta('pinned');
   @override
   late final GeneratedColumn<bool> pinned = GeneratedColumn<bool>(
@@ -920,6 +926,7 @@ class $MessagesTable extends Messages
         remoteDeletedAt,
         messageTextUpdatedAt,
         userId,
+        channelRole,
         pinned,
         pinnedAt,
         pinExpires,
@@ -1040,6 +1047,12 @@ class $MessagesTable extends Messages
       context.handle(_userIdMeta,
           userId.isAcceptableOrUnknown(data['user_id']!, _userIdMeta));
     }
+    if (data.containsKey('channel_role')) {
+      context.handle(
+          _channelRoleMeta,
+          channelRole.isAcceptableOrUnknown(
+              data['channel_role']!, _channelRoleMeta));
+    }
     if (data.containsKey('pinned')) {
       context.handle(_pinnedMeta,
           pinned.isAcceptableOrUnknown(data['pinned']!, _pinnedMeta));
@@ -1125,6 +1138,8 @@ class $MessagesTable extends Messages
           data['${effectivePrefix}message_text_updated_at']),
       userId: attachedDatabase.typeMapping
           .read(DriftSqlType.string, data['${effectivePrefix}user_id']),
+      channelRole: attachedDatabase.typeMapping
+          .read(DriftSqlType.string, data['${effectivePrefix}channel_role']),
       pinned: attachedDatabase.typeMapping
           .read(DriftSqlType.bool, data['${effectivePrefix}pinned'])!,
       pinnedAt: attachedDatabase.typeMapping
@@ -1240,6 +1255,9 @@ class MessageEntity extends DataClass implements Insertable<MessageEntity> {
   /// Id of the User who sent the message
   final String? userId;
 
+  /// The channel role of the user who sent the message
+  final String? channelRole;
+
   /// Whether the message is pinned or not
   final bool pinned;
 
@@ -1286,6 +1304,7 @@ class MessageEntity extends DataClass implements Insertable<MessageEntity> {
       this.remoteDeletedAt,
       this.messageTextUpdatedAt,
       this.userId,
+      this.channelRole,
       required this.pinned,
       this.pinnedAt,
       this.pinExpires,
@@ -1358,6 +1377,9 @@ class MessageEntity extends DataClass implements Insertable<MessageEntity> {
     if (!nullToAbsent || userId != null) {
       map['user_id'] = Variable<String>(userId);
     }
+    if (!nullToAbsent || channelRole != null) {
+      map['channel_role'] = Variable<String>(channelRole);
+    }
     map['pinned'] = Variable<bool>(pinned);
     if (!nullToAbsent || pinnedAt != null) {
       map['pinned_at'] = Variable<DateTime>(pinnedAt);
@@ -1412,6 +1434,7 @@ class MessageEntity extends DataClass implements Insertable<MessageEntity> {
       messageTextUpdatedAt:
           serializer.fromJson<DateTime?>(json['messageTextUpdatedAt']),
       userId: serializer.fromJson<String?>(json['userId']),
+      channelRole: serializer.fromJson<String?>(json['channelRole']),
       pinned: serializer.fromJson<bool>(json['pinned']),
       pinnedAt: serializer.fromJson<DateTime?>(json['pinnedAt']),
       pinExpires: serializer.fromJson<DateTime?>(json['pinExpires']),
@@ -1451,6 +1474,7 @@ class MessageEntity extends DataClass implements Insertable<MessageEntity> {
       'messageTextUpdatedAt':
           serializer.toJson<DateTime?>(messageTextUpdatedAt),
       'userId': serializer.toJson<String?>(userId),
+      'channelRole': serializer.toJson<String?>(channelRole),
       'pinned': serializer.toJson<bool>(pinned),
       'pinnedAt': serializer.toJson<DateTime?>(pinnedAt),
       'pinExpires': serializer.toJson<DateTime?>(pinExpires),
@@ -1487,6 +1511,7 @@ class MessageEntity extends DataClass implements Insertable<MessageEntity> {
           Value<DateTime?> remoteDeletedAt = const Value.absent(),
           Value<DateTime?> messageTextUpdatedAt = const Value.absent(),
           Value<String?> userId = const Value.absent(),
+          Value<String?> channelRole = const Value.absent(),
           bool? pinned,
           Value<DateTime?> pinnedAt = const Value.absent(),
           Value<DateTime?> pinExpires = const Value.absent(),
@@ -1533,6 +1558,7 @@ class MessageEntity extends DataClass implements Insertable<MessageEntity> {
             ? messageTextUpdatedAt.value
             : this.messageTextUpdatedAt,
         userId: userId.present ? userId.value : this.userId,
+        channelRole: channelRole.present ? channelRole.value : this.channelRole,
         pinned: pinned ?? this.pinned,
         pinnedAt: pinnedAt.present ? pinnedAt.value : this.pinnedAt,
         pinExpires: pinExpires.present ? pinExpires.value : this.pinExpires,
@@ -1594,6 +1620,8 @@ class MessageEntity extends DataClass implements Insertable<MessageEntity> {
           ? data.messageTextUpdatedAt.value
           : this.messageTextUpdatedAt,
       userId: data.userId.present ? data.userId.value : this.userId,
+      channelRole:
+          data.channelRole.present ? data.channelRole.value : this.channelRole,
       pinned: data.pinned.present ? data.pinned.value : this.pinned,
       pinnedAt: data.pinnedAt.present ? data.pinnedAt.value : this.pinnedAt,
       pinExpires:
@@ -1636,6 +1664,7 @@ class MessageEntity extends DataClass implements Insertable<MessageEntity> {
           ..write('remoteDeletedAt: $remoteDeletedAt, ')
           ..write('messageTextUpdatedAt: $messageTextUpdatedAt, ')
           ..write('userId: $userId, ')
+          ..write('channelRole: $channelRole, ')
           ..write('pinned: $pinned, ')
           ..write('pinnedAt: $pinnedAt, ')
           ..write('pinExpires: $pinExpires, ')
@@ -1672,6 +1701,7 @@ class MessageEntity extends DataClass implements Insertable<MessageEntity> {
         remoteDeletedAt,
         messageTextUpdatedAt,
         userId,
+        channelRole,
         pinned,
         pinnedAt,
         pinExpires,
@@ -1707,6 +1737,7 @@ class MessageEntity extends DataClass implements Insertable<MessageEntity> {
           other.remoteDeletedAt == this.remoteDeletedAt &&
           other.messageTextUpdatedAt == this.messageTextUpdatedAt &&
           other.userId == this.userId &&
+          other.channelRole == this.channelRole &&
           other.pinned == this.pinned &&
           other.pinnedAt == this.pinnedAt &&
           other.pinExpires == this.pinExpires &&
@@ -1740,6 +1771,7 @@ class MessagesCompanion extends UpdateCompanion<MessageEntity> {
   final Value<DateTime?> remoteDeletedAt;
   final Value<DateTime?> messageTextUpdatedAt;
   final Value<String?> userId;
+  final Value<String?> channelRole;
   final Value<bool> pinned;
   final Value<DateTime?> pinnedAt;
   final Value<DateTime?> pinExpires;
@@ -1772,6 +1804,7 @@ class MessagesCompanion extends UpdateCompanion<MessageEntity> {
     this.remoteDeletedAt = const Value.absent(),
     this.messageTextUpdatedAt = const Value.absent(),
     this.userId = const Value.absent(),
+    this.channelRole = const Value.absent(),
     this.pinned = const Value.absent(),
     this.pinnedAt = const Value.absent(),
     this.pinExpires = const Value.absent(),
@@ -1805,6 +1838,7 @@ class MessagesCompanion extends UpdateCompanion<MessageEntity> {
     this.remoteDeletedAt = const Value.absent(),
     this.messageTextUpdatedAt = const Value.absent(),
     this.userId = const Value.absent(),
+    this.channelRole = const Value.absent(),
     this.pinned = const Value.absent(),
     this.pinnedAt = const Value.absent(),
     this.pinExpires = const Value.absent(),
@@ -1842,6 +1876,7 @@ class MessagesCompanion extends UpdateCompanion<MessageEntity> {
     Expression<DateTime>? remoteDeletedAt,
     Expression<DateTime>? messageTextUpdatedAt,
     Expression<String>? userId,
+    Expression<String>? channelRole,
     Expression<bool>? pinned,
     Expression<DateTime>? pinnedAt,
     Expression<DateTime>? pinExpires,
@@ -1876,6 +1911,7 @@ class MessagesCompanion extends UpdateCompanion<MessageEntity> {
       if (messageTextUpdatedAt != null)
         'message_text_updated_at': messageTextUpdatedAt,
       if (userId != null) 'user_id': userId,
+      if (channelRole != null) 'channel_role': channelRole,
       if (pinned != null) 'pinned': pinned,
       if (pinnedAt != null) 'pinned_at': pinnedAt,
       if (pinExpires != null) 'pin_expires': pinExpires,
@@ -1912,6 +1948,7 @@ class MessagesCompanion extends UpdateCompanion<MessageEntity> {
       Value<DateTime?>? remoteDeletedAt,
       Value<DateTime?>? messageTextUpdatedAt,
       Value<String?>? userId,
+      Value<String?>? channelRole,
       Value<bool>? pinned,
       Value<DateTime?>? pinnedAt,
       Value<DateTime?>? pinExpires,
@@ -1944,6 +1981,7 @@ class MessagesCompanion extends UpdateCompanion<MessageEntity> {
       remoteDeletedAt: remoteDeletedAt ?? this.remoteDeletedAt,
       messageTextUpdatedAt: messageTextUpdatedAt ?? this.messageTextUpdatedAt,
       userId: userId ?? this.userId,
+      channelRole: channelRole ?? this.channelRole,
       pinned: pinned ?? this.pinned,
       pinnedAt: pinnedAt ?? this.pinnedAt,
       pinExpires: pinExpires ?? this.pinExpires,
@@ -2029,6 +2067,9 @@ class MessagesCompanion extends UpdateCompanion<MessageEntity> {
     if (userId.present) {
       map['user_id'] = Variable<String>(userId.value);
     }
+    if (channelRole.present) {
+      map['channel_role'] = Variable<String>(channelRole.value);
+    }
     if (pinned.present) {
       map['pinned'] = Variable<bool>(pinned.value);
     }
@@ -2088,6 +2129,7 @@ class MessagesCompanion extends UpdateCompanion<MessageEntity> {
           ..write('remoteDeletedAt: $remoteDeletedAt, ')
           ..write('messageTextUpdatedAt: $messageTextUpdatedAt, ')
           ..write('userId: $userId, ')
+          ..write('channelRole: $channelRole, ')
           ..write('pinned: $pinned, ')
           ..write('pinnedAt: $pinnedAt, ')
           ..write('pinExpires: $pinExpires, ')
@@ -2952,6 +2994,12 @@ class $PinnedMessagesTable extends PinnedMessages
   late final GeneratedColumn<String> userId = GeneratedColumn<String>(
       'user_id', aliasedName, true,
       type: DriftSqlType.string, requiredDuringInsert: false);
+  static const VerificationMeta _channelRoleMeta =
+      const VerificationMeta('channelRole');
+  @override
+  late final GeneratedColumn<String> channelRole = GeneratedColumn<String>(
+      'channel_role', aliasedName, true,
+      type: DriftSqlType.string, requiredDuringInsert: false);
   static const VerificationMeta _pinnedMeta = const VerificationMeta('pinned');
   @override
   late final GeneratedColumn<bool> pinned = GeneratedColumn<bool>(
@@ -3028,6 +3076,7 @@ class $PinnedMessagesTable extends PinnedMessages
         remoteDeletedAt,
         messageTextUpdatedAt,
         userId,
+        channelRole,
         pinned,
         pinnedAt,
         pinExpires,
@@ -3149,6 +3198,12 @@ class $PinnedMessagesTable extends PinnedMessages
       context.handle(_userIdMeta,
           userId.isAcceptableOrUnknown(data['user_id']!, _userIdMeta));
     }
+    if (data.containsKey('channel_role')) {
+      context.handle(
+          _channelRoleMeta,
+          channelRole.isAcceptableOrUnknown(
+              data['channel_role']!, _channelRoleMeta));
+    }
     if (data.containsKey('pinned')) {
       context.handle(_pinnedMeta,
           pinned.isAcceptableOrUnknown(data['pinned']!, _pinnedMeta));
@@ -3234,6 +3289,8 @@ class $PinnedMessagesTable extends PinnedMessages
           data['${effectivePrefix}message_text_updated_at']),
       userId: attachedDatabase.typeMapping
           .read(DriftSqlType.string, data['${effectivePrefix}user_id']),
+      channelRole: attachedDatabase.typeMapping
+          .read(DriftSqlType.string, data['${effectivePrefix}channel_role']),
       pinned: attachedDatabase.typeMapping
           .read(DriftSqlType.bool, data['${effectivePrefix}pinned'])!,
       pinnedAt: attachedDatabase.typeMapping
@@ -3351,6 +3408,9 @@ class PinnedMessageEntity extends DataClass
   /// Id of the User who sent the message
   final String? userId;
 
+  /// The channel role of the user who sent the message
+  final String? channelRole;
+
   /// Whether the message is pinned or not
   final bool pinned;
 
@@ -3397,6 +3457,7 @@ class PinnedMessageEntity extends DataClass
       this.remoteDeletedAt,
       this.messageTextUpdatedAt,
       this.userId,
+      this.channelRole,
       required this.pinned,
       this.pinnedAt,
       this.pinExpires,
@@ -3469,6 +3530,9 @@ class PinnedMessageEntity extends DataClass
     if (!nullToAbsent || userId != null) {
       map['user_id'] = Variable<String>(userId);
     }
+    if (!nullToAbsent || channelRole != null) {
+      map['channel_role'] = Variable<String>(channelRole);
+    }
     map['pinned'] = Variable<bool>(pinned);
     if (!nullToAbsent || pinnedAt != null) {
       map['pinned_at'] = Variable<DateTime>(pinnedAt);
@@ -3524,6 +3588,7 @@ class PinnedMessageEntity extends DataClass
       messageTextUpdatedAt:
           serializer.fromJson<DateTime?>(json['messageTextUpdatedAt']),
       userId: serializer.fromJson<String?>(json['userId']),
+      channelRole: serializer.fromJson<String?>(json['channelRole']),
       pinned: serializer.fromJson<bool>(json['pinned']),
       pinnedAt: serializer.fromJson<DateTime?>(json['pinnedAt']),
       pinExpires: serializer.fromJson<DateTime?>(json['pinExpires']),
@@ -3563,6 +3628,7 @@ class PinnedMessageEntity extends DataClass
       'messageTextUpdatedAt':
           serializer.toJson<DateTime?>(messageTextUpdatedAt),
       'userId': serializer.toJson<String?>(userId),
+      'channelRole': serializer.toJson<String?>(channelRole),
       'pinned': serializer.toJson<bool>(pinned),
       'pinnedAt': serializer.toJson<DateTime?>(pinnedAt),
       'pinExpires': serializer.toJson<DateTime?>(pinExpires),
@@ -3599,6 +3665,7 @@ class PinnedMessageEntity extends DataClass
           Value<DateTime?> remoteDeletedAt = const Value.absent(),
           Value<DateTime?> messageTextUpdatedAt = const Value.absent(),
           Value<String?> userId = const Value.absent(),
+          Value<String?> channelRole = const Value.absent(),
           bool? pinned,
           Value<DateTime?> pinnedAt = const Value.absent(),
           Value<DateTime?> pinExpires = const Value.absent(),
@@ -3645,6 +3712,7 @@ class PinnedMessageEntity extends DataClass
             ? messageTextUpdatedAt.value
             : this.messageTextUpdatedAt,
         userId: userId.present ? userId.value : this.userId,
+        channelRole: channelRole.present ? channelRole.value : this.channelRole,
         pinned: pinned ?? this.pinned,
         pinnedAt: pinnedAt.present ? pinnedAt.value : this.pinnedAt,
         pinExpires: pinExpires.present ? pinExpires.value : this.pinExpires,
@@ -3706,6 +3774,8 @@ class PinnedMessageEntity extends DataClass
           ? data.messageTextUpdatedAt.value
           : this.messageTextUpdatedAt,
       userId: data.userId.present ? data.userId.value : this.userId,
+      channelRole:
+          data.channelRole.present ? data.channelRole.value : this.channelRole,
       pinned: data.pinned.present ? data.pinned.value : this.pinned,
       pinnedAt: data.pinnedAt.present ? data.pinnedAt.value : this.pinnedAt,
       pinExpires:
@@ -3748,6 +3818,7 @@ class PinnedMessageEntity extends DataClass
           ..write('remoteDeletedAt: $remoteDeletedAt, ')
           ..write('messageTextUpdatedAt: $messageTextUpdatedAt, ')
           ..write('userId: $userId, ')
+          ..write('channelRole: $channelRole, ')
           ..write('pinned: $pinned, ')
           ..write('pinnedAt: $pinnedAt, ')
           ..write('pinExpires: $pinExpires, ')
@@ -3784,6 +3855,7 @@ class PinnedMessageEntity extends DataClass
         remoteDeletedAt,
         messageTextUpdatedAt,
         userId,
+        channelRole,
         pinned,
         pinnedAt,
         pinExpires,
@@ -3819,6 +3891,7 @@ class PinnedMessageEntity extends DataClass
           other.remoteDeletedAt == this.remoteDeletedAt &&
           other.messageTextUpdatedAt == this.messageTextUpdatedAt &&
           other.userId == this.userId &&
+          other.channelRole == this.channelRole &&
           other.pinned == this.pinned &&
           other.pinnedAt == this.pinnedAt &&
           other.pinExpires == this.pinExpires &&
@@ -3852,6 +3925,7 @@ class PinnedMessagesCompanion extends UpdateCompanion<PinnedMessageEntity> {
   final Value<DateTime?> remoteDeletedAt;
   final Value<DateTime?> messageTextUpdatedAt;
   final Value<String?> userId;
+  final Value<String?> channelRole;
   final Value<bool> pinned;
   final Value<DateTime?> pinnedAt;
   final Value<DateTime?> pinExpires;
@@ -3884,6 +3958,7 @@ class PinnedMessagesCompanion extends UpdateCompanion<PinnedMessageEntity> {
     this.remoteDeletedAt = const Value.absent(),
     this.messageTextUpdatedAt = const Value.absent(),
     this.userId = const Value.absent(),
+    this.channelRole = const Value.absent(),
     this.pinned = const Value.absent(),
     this.pinnedAt = const Value.absent(),
     this.pinExpires = const Value.absent(),
@@ -3917,6 +3992,7 @@ class PinnedMessagesCompanion extends UpdateCompanion<PinnedMessageEntity> {
     this.remoteDeletedAt = const Value.absent(),
     this.messageTextUpdatedAt = const Value.absent(),
     this.userId = const Value.absent(),
+    this.channelRole = const Value.absent(),
     this.pinned = const Value.absent(),
     this.pinnedAt = const Value.absent(),
     this.pinExpires = const Value.absent(),
@@ -3954,6 +4030,7 @@ class PinnedMessagesCompanion extends UpdateCompanion<PinnedMessageEntity> {
     Expression<DateTime>? remoteDeletedAt,
     Expression<DateTime>? messageTextUpdatedAt,
     Expression<String>? userId,
+    Expression<String>? channelRole,
     Expression<bool>? pinned,
     Expression<DateTime>? pinnedAt,
     Expression<DateTime>? pinExpires,
@@ -3988,6 +4065,7 @@ class PinnedMessagesCompanion extends UpdateCompanion<PinnedMessageEntity> {
       if (messageTextUpdatedAt != null)
         'message_text_updated_at': messageTextUpdatedAt,
       if (userId != null) 'user_id': userId,
+      if (channelRole != null) 'channel_role': channelRole,
       if (pinned != null) 'pinned': pinned,
       if (pinnedAt != null) 'pinned_at': pinnedAt,
       if (pinExpires != null) 'pin_expires': pinExpires,
@@ -4024,6 +4102,7 @@ class PinnedMessagesCompanion extends UpdateCompanion<PinnedMessageEntity> {
       Value<DateTime?>? remoteDeletedAt,
       Value<DateTime?>? messageTextUpdatedAt,
       Value<String?>? userId,
+      Value<String?>? channelRole,
       Value<bool>? pinned,
       Value<DateTime?>? pinnedAt,
       Value<DateTime?>? pinExpires,
@@ -4056,6 +4135,7 @@ class PinnedMessagesCompanion extends UpdateCompanion<PinnedMessageEntity> {
       remoteDeletedAt: remoteDeletedAt ?? this.remoteDeletedAt,
       messageTextUpdatedAt: messageTextUpdatedAt ?? this.messageTextUpdatedAt,
       userId: userId ?? this.userId,
+      channelRole: channelRole ?? this.channelRole,
       pinned: pinned ?? this.pinned,
       pinnedAt: pinnedAt ?? this.pinnedAt,
       pinExpires: pinExpires ?? this.pinExpires,
@@ -4143,6 +4223,9 @@ class PinnedMessagesCompanion extends UpdateCompanion<PinnedMessageEntity> {
     if (userId.present) {
       map['user_id'] = Variable<String>(userId.value);
     }
+    if (channelRole.present) {
+      map['channel_role'] = Variable<String>(channelRole.value);
+    }
     if (pinned.present) {
       map['pinned'] = Variable<bool>(pinned.value);
     }
@@ -4202,6 +4285,7 @@ class PinnedMessagesCompanion extends UpdateCompanion<PinnedMessageEntity> {
           ..write('remoteDeletedAt: $remoteDeletedAt, ')
           ..write('messageTextUpdatedAt: $messageTextUpdatedAt, ')
           ..write('userId: $userId, ')
+          ..write('channelRole: $channelRole, ')
           ..write('pinned: $pinned, ')
           ..write('pinnedAt: $pinnedAt, ')
           ..write('pinExpires: $pinExpires, ')
@@ -9116,6 +9200,7 @@ typedef $$MessagesTableCreateCompanionBuilder = MessagesCompanion Function({
   Value<DateTime?> remoteDeletedAt,
   Value<DateTime?> messageTextUpdatedAt,
   Value<String?> userId,
+  Value<String?> channelRole,
   Value<bool> pinned,
   Value<DateTime?> pinnedAt,
   Value<DateTime?> pinExpires,
@@ -9149,6 +9234,7 @@ typedef $$MessagesTableUpdateCompanionBuilder = MessagesCompanion Function({
   Value<DateTime?> remoteDeletedAt,
   Value<DateTime?> messageTextUpdatedAt,
   Value<String?> userId,
+  Value<String?> channelRole,
   Value<bool> pinned,
   Value<DateTime?> pinnedAt,
   Value<DateTime?> pinExpires,
@@ -9299,6 +9385,9 @@ class $$MessagesTableFilterComposer
 
   ColumnFilters<String> get userId => $composableBuilder(
       column: $table.userId, builder: (column) => ColumnFilters(column));
+
+  ColumnFilters<String> get channelRole => $composableBuilder(
+      column: $table.channelRole, builder: (column) => ColumnFilters(column));
 
   ColumnFilters<bool> get pinned => $composableBuilder(
       column: $table.pinned, builder: (column) => ColumnFilters(column));
@@ -9479,6 +9568,9 @@ class $$MessagesTableOrderingComposer
   ColumnOrderings<String> get userId => $composableBuilder(
       column: $table.userId, builder: (column) => ColumnOrderings(column));
 
+  ColumnOrderings<String> get channelRole => $composableBuilder(
+      column: $table.channelRole, builder: (column) => ColumnOrderings(column));
+
   ColumnOrderings<bool> get pinned => $composableBuilder(
       column: $table.pinned, builder: (column) => ColumnOrderings(column));
 
@@ -9600,6 +9692,9 @@ class $$MessagesTableAnnotationComposer
 
   GeneratedColumn<String> get userId =>
       $composableBuilder(column: $table.userId, builder: (column) => column);
+
+  GeneratedColumn<String> get channelRole => $composableBuilder(
+      column: $table.channelRole, builder: (column) => column);
 
   GeneratedColumn<bool> get pinned =>
       $composableBuilder(column: $table.pinned, builder: (column) => column);
@@ -9734,6 +9829,7 @@ class $$MessagesTableTableManager extends RootTableManager<
             Value<DateTime?> remoteDeletedAt = const Value.absent(),
             Value<DateTime?> messageTextUpdatedAt = const Value.absent(),
             Value<String?> userId = const Value.absent(),
+            Value<String?> channelRole = const Value.absent(),
             Value<bool> pinned = const Value.absent(),
             Value<DateTime?> pinnedAt = const Value.absent(),
             Value<DateTime?> pinExpires = const Value.absent(),
@@ -9767,6 +9863,7 @@ class $$MessagesTableTableManager extends RootTableManager<
             remoteDeletedAt: remoteDeletedAt,
             messageTextUpdatedAt: messageTextUpdatedAt,
             userId: userId,
+            channelRole: channelRole,
             pinned: pinned,
             pinnedAt: pinnedAt,
             pinExpires: pinExpires,
@@ -9801,6 +9898,7 @@ class $$MessagesTableTableManager extends RootTableManager<
             Value<DateTime?> remoteDeletedAt = const Value.absent(),
             Value<DateTime?> messageTextUpdatedAt = const Value.absent(),
             Value<String?> userId = const Value.absent(),
+            Value<String?> channelRole = const Value.absent(),
             Value<bool> pinned = const Value.absent(),
             Value<DateTime?> pinnedAt = const Value.absent(),
             Value<DateTime?> pinExpires = const Value.absent(),
@@ -9834,6 +9932,7 @@ class $$MessagesTableTableManager extends RootTableManager<
             remoteDeletedAt: remoteDeletedAt,
             messageTextUpdatedAt: messageTextUpdatedAt,
             userId: userId,
+            channelRole: channelRole,
             pinned: pinned,
             pinnedAt: pinnedAt,
             pinExpires: pinExpires,
@@ -10458,6 +10557,7 @@ typedef $$PinnedMessagesTableCreateCompanionBuilder = PinnedMessagesCompanion
   Value<DateTime?> remoteDeletedAt,
   Value<DateTime?> messageTextUpdatedAt,
   Value<String?> userId,
+  Value<String?> channelRole,
   Value<bool> pinned,
   Value<DateTime?> pinnedAt,
   Value<DateTime?> pinExpires,
@@ -10492,6 +10592,7 @@ typedef $$PinnedMessagesTableUpdateCompanionBuilder = PinnedMessagesCompanion
   Value<DateTime?> remoteDeletedAt,
   Value<DateTime?> messageTextUpdatedAt,
   Value<String?> userId,
+  Value<String?> channelRole,
   Value<bool> pinned,
   Value<DateTime?> pinnedAt,
   Value<DateTime?> pinExpires,
@@ -10617,6 +10718,9 @@ class $$PinnedMessagesTableFilterComposer
 
   ColumnFilters<String> get userId => $composableBuilder(
       column: $table.userId, builder: (column) => ColumnFilters(column));
+
+  ColumnFilters<String> get channelRole => $composableBuilder(
+      column: $table.channelRole, builder: (column) => ColumnFilters(column));
 
   ColumnFilters<bool> get pinned => $composableBuilder(
       column: $table.pinned, builder: (column) => ColumnFilters(column));
@@ -10761,6 +10865,9 @@ class $$PinnedMessagesTableOrderingComposer
   ColumnOrderings<String> get userId => $composableBuilder(
       column: $table.userId, builder: (column) => ColumnOrderings(column));
 
+  ColumnOrderings<String> get channelRole => $composableBuilder(
+      column: $table.channelRole, builder: (column) => ColumnOrderings(column));
+
   ColumnOrderings<bool> get pinned => $composableBuilder(
       column: $table.pinned, builder: (column) => ColumnOrderings(column));
 
@@ -10866,6 +10973,9 @@ class $$PinnedMessagesTableAnnotationComposer
   GeneratedColumn<String> get userId =>
       $composableBuilder(column: $table.userId, builder: (column) => column);
 
+  GeneratedColumn<String> get channelRole => $composableBuilder(
+      column: $table.channelRole, builder: (column) => column);
+
   GeneratedColumn<bool> get pinned =>
       $composableBuilder(column: $table.pinned, builder: (column) => column);
 
@@ -10963,6 +11073,7 @@ class $$PinnedMessagesTableTableManager extends RootTableManager<
             Value<DateTime?> remoteDeletedAt = const Value.absent(),
             Value<DateTime?> messageTextUpdatedAt = const Value.absent(),
             Value<String?> userId = const Value.absent(),
+            Value<String?> channelRole = const Value.absent(),
             Value<bool> pinned = const Value.absent(),
             Value<DateTime?> pinnedAt = const Value.absent(),
             Value<DateTime?> pinExpires = const Value.absent(),
@@ -10996,6 +11107,7 @@ class $$PinnedMessagesTableTableManager extends RootTableManager<
             remoteDeletedAt: remoteDeletedAt,
             messageTextUpdatedAt: messageTextUpdatedAt,
             userId: userId,
+            channelRole: channelRole,
             pinned: pinned,
             pinnedAt: pinnedAt,
             pinExpires: pinExpires,
@@ -11030,6 +11142,7 @@ class $$PinnedMessagesTableTableManager extends RootTableManager<
             Value<DateTime?> remoteDeletedAt = const Value.absent(),
             Value<DateTime?> messageTextUpdatedAt = const Value.absent(),
             Value<String?> userId = const Value.absent(),
+            Value<String?> channelRole = const Value.absent(),
             Value<bool> pinned = const Value.absent(),
             Value<DateTime?> pinnedAt = const Value.absent(),
             Value<DateTime?> pinExpires = const Value.absent(),
@@ -11063,6 +11176,7 @@ class $$PinnedMessagesTableTableManager extends RootTableManager<
             remoteDeletedAt: remoteDeletedAt,
             messageTextUpdatedAt: messageTextUpdatedAt,
             userId: userId,
+            channelRole: channelRole,
             pinned: pinned,
             pinnedAt: pinnedAt,
             pinExpires: pinExpires,